--- index.js
@@ -1,3 +1,0 @@
-const ct = require('./src');
-
-exports.module = { ct };
