+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@noob-zeroo/chrono-trigger",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Library for measuring the performance of a code trail",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
