+++ src/index.js
@@ -1,16 +1,2 @@
-const { performance } = require('perf_hooks');
-const times = {};
-
-module.exports = (label) => {
-    const now = performance.now();
-
-    if (!times[label]) {
-        times[label] = now;
-        return;
-    }
-
-    const beginTime = times[label];
-    const elapsed = now - beginTime;
-    delete times[label];
-    return { label, elapsed: Number.parseFloat(elapsed.toFixed(2)) };
-}
+const ct = require('ct');
+module.exports = { ct };
